third_party/blink/web_tests/fast/dom/shadow/parent-tree-scope-in-shadow-expected.txt

CONSOLE WARNING: line 14: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Creating shadow dom subtrees from top to bottom.
PASS internals.parentTreeScope(shadow1) is document
PASS internals.parentTreeScope(shadow2) is shadow1
PASS internals.parentTreeScope(shadow3) is shadow2
Creating multiple shadow dom subtrees from bottom to top.
PASS internals.parentTreeScope(shadow1) is document
PASS internals.parentTreeScope(shadow2) is shadow1
PASS internals.parentTreeScope(shadow3) is shadow2
PASS successfullyParsed is true

TEST COMPLETE
